Extract cover image and store in has_part
During the publisher backlist ingest, extract cover images and store
them in the Record `has_part` field as a cover, which later gets
attached to the Edition as a link during clustering.

Note, I added a bit of refactoring:

  1. Parse out the hathi id outside of `download_file_location` which
     reduces nesting in that method
  2. Use `self.file_bucket` instead of directly accessing `os` method

I tried to split this up a bit, but it proved cumbersome. Also note,
this is kinda slow