Generate covers during publisher backlist ingest (#670)
* Add a PDF cover generator

Add a simple class that reads in a bytestream into a PDF, and provides a
simple method for inferring the cover image from the PDF.

* Extract cover image and store in has_part

During the publisher backlist ingest, extract cover images and store
them in the Record `has_part` field as a cover, which later gets
attached to the Edition as a link during clustering.

Note, I added a bit of refactoring:

  1. Parse out the hathi id outside of `download_file_location` which
     reduces nesting in that method
  2. Use `self.file_bucket` instead of directly accessing `os` method

I tried to split this up a bit, but it proved cumbersome. Also note,
this is kinda slow

* Split out method call into lines

Just for a little readability

* Fix get_hathi_id typing

It's a string and python3.9 doesn't support `|` for union types

* Set cover media_type to image/png

Need this media type to make sure this image gets surfaced in opds