Eager load work and ids in collections queries (#676)
* Eager load work and ids in collections queries

Looking at some of the slowness in newrelic, noticed that we're issuing
500 work queries and 500 identifier queries per request, due to fetching
works and ids for each edition in every collection.

To alleviate this, just fetch everything upfront, using `joinedload` as
we do for other entities.  I did a bit of refactoring to make both
collection queries (ie, a feed vs a single colleciton) use the same base
query.  I also had to add a bit of sauce to make sorting work, as the
addition of work join caused column name ambiguity in the sort we were
doing.

* Handle default sort on fetchCollections

Preserve the default of title sort if no sort is passed in

* Remove import

Didn't end up using this

* Formatting

Add some indentaiton and joinedload the Work with an extra param instead
of chaining to match the existing pattern