fix: Properly load covers on collection page
Was testing some of the server side cover work I did and noticed I could
get covers to load for an individual edition but not on the collection
page - inspected the components and looks like we were trying to pass an
OpdsPublication to a cover method that expects a full feed. Easy enough
to just write a new cover finding method for a single edition.