Directly add covers to Schomburg editions (#687)
There's some order of operation issues wrt the Schomburg ingest and
the covers - since the ingest was done before the cover stuff was added
to it, the editions didn't get cover images and becaue of the
non-determinism of the clustering, I don't want to re-run the ingest, as
it will affect the number of editions / clusters in the collection.

So instead, just add the covers directly onto the images.