Finish QA deploy before integration tests run (#690)
Something I noticed the other day - we tell ECS to deploy the new image
and then immediately run integration tests. However, with how ECS
deployments work, it's not clear whether the tests are hitting the old
or new tasks. Waiting for services to be stable will ensure we're
testing against the new image and will also avoid running the tests if
the new instance never becomes healthy