Migrate gutenberg to RecordIngestor
Migrate to using the record ingestor for the gutenberg ingest. Tracing
through this, I noticed that the mapping creates some record parts, but
then the ingest clears them out and reworks them to add the s3 urls.
Instead, I opted to just move everything into the ingest, but I can
instead invert that and move it all back to the mapping.