Decouple RecordIngestor from SourceService
Some ingests have custom manifest logic, and future ingest processes
might have more custom processing. So decouple the source service from
the ingestor, instead simply passing a (usually lazy) collection of
records to the ingestor, allowing the ingest process to do any special
processing of records before handing them off to the ingestor.

For now, the changes in this PR are boilerplate, since we haven't
transitioned any of the special ingests to the record ingestor