Migrate gutenberg to RecordIngestor (#701)
* Migrate gutenberg to RecordIngestor

Migrate to using the record ingestor for the gutenberg ingest. Tracing
through this, I noticed that the mapping creates some record parts, but
then the ingest clears them out and reworks them to add the s3 urls.
Instead, I opted to just move everything into the ingest, but I can
instead invert that and move it all back to the mapping.

* Use a mock webpub converter

Add a simple fixture that just returns a locally stored manifest file in
place of hitting the webpub converter, since the remote converter
can't talk to local s3

* fix: Fix webpub part content type

* Check for cover existence

* Swap to `get`

The `head` requests seems to return 302s for everything :shrug

* fix: Fix gutenberg service pagination

Realized that we mutate `current_position` while checking for the
offset, which results in us yielding too many records in the inner loop.
Instead, move the current_position update into the inner loop, and allow
the inner loop to break and exit the outer loop upon reaching the limit.

* Follow redirects for cover check

Revert back to a head request, but follow redirects, since it seems a
lot of these covers give back a 302

* Move has_part logic to mapping

* fix: Fix gutenberg mapping unit test

* Use record source for has_part

* fix: Record.source is already a string

* more retries

:shrug

* fix test

bad merge smh

* fix: Stop parallelizing tests

Seems to lead to conflicts between processes talking to the queue

* Skip publisher backlist test

It appears to be flaky

* Mock epub to webpub in loc ingest test