Persist records and set state in ingestor
when ingesting records, persist them with a new `state` value of
`ingested`, and then send a smaller message to SQS, just containing the
record id to stay under the SQS message size limit.

Note, I adapted the record buffer a bit here so that the ingestor could
consume the buffered records as they get flushed out - I *could* instead
replicate / extract the upsert logic in the record buffer into the
ingestor, but this way was fastest.