Add execution time to success event
Add a simple `perf_counter` based implementation to update the execution
time of successful messages. I could also add this to the failed
message, but it seems like this could be not super useful depending on
the nature of the failures.

I think a better thing might be to add execution times to external
service requests. To that end, I can implement some kind of decorator /
context manager to make this easy to port, but figured I'd start simple
until we agree on approach.