Add RecordPipeline success / failure events  (#733)
* add Monitor service

* track WorkRecordsChosen

* switch to non-class Monitor

* upgrade newrelic agent

* add explicit agent shutdown

* add work record tracking

* remove unused import

* update cluster tracking

* add more events

* track record embellisher

* remove logs

* change import style

* track editions

* change import casing of `monitor`

* remove test newrelic script

* fix record finder newrelic call

* Add RecordPipeline success / failure events

Building off @kristojorg's monitoring work, add a couple more events for
tracking successful and failed messages in the RecordPipeline.

* Add execution time to success event

Add a simple `perf_counter` based implementation to update the execution
time of successful messages. I could also add this to the failed
message, but it seems like this could be not super useful depending on
the nature of the failures.

I think a better thing might be to add execution times to external
service requests. To that end, I can implement some kind of decorator /
context manager to make this easy to port, but figured I'd start simple
until we agree on approach.

* Add execution time to failed events

Easy enough to track :shrug

---------

Co-authored-by: Kristo Jorgenson <kristo.jorgenson@gmail.com>