refactor to read log level from env at execution time
relies on python logging hierarchy to set a root logger
for the entire application and configure the application
root logger before execution.

This follows the principle that env vars should be read
at execution and not during import. Previously
LOG LEVEL was read from env var on import of any
process module.

Risks: application monitoring queries of logs DB might
break with the addition of 'drb.' to the start of
every logger name.