use environment in retrieved ssm parameters
The previous code state required part of the name of any
retrieved SSM parameter store parameter to be either
"qa" or "production" regardless of the actual value
of the "ENVIRONMENT" env variable. During test execution
in GHA the "ENVIRONMENT" env var value is "local" not
"qa" or "production". The result was confusion of about
the semantics of the ENVIRONMENT env var and its
relationship to the secrets used. Was ENVIRONMENT
meant to determine which secrets are used or not?
The previous code state allowed the relationship btw
ENVIRONMENT and the actual parameter value used to
be opaque and hidden in this implicit code.